src/ai_modules.py
import torch
import torch.nn as nn
import torchvision.models as models
import numpy as np
import cv2
from PIL import Image
import tensorflow as tf
from tensorflow.keras.applications import VGG16, ResNet50
from tensorflow.keras.models import Model

class SuperResolution:
    def __init__(self):
        # SRGAN benzeri bir model kullanacağız
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.model = models.resnet34(pretrained=True)
        self.model = self.model.to(self.device)
        
    def upscale(self, image, scale_factor=2):
        """Görüntü çözünürlüğünü artır"""
        img_array = np.array(image)
        # Görüntüyü model için hazırla
        input_tensor = torch.from_numpy(img_array).float().permute(2, 0, 1).unsqueeze(0)
        input_tensor = input_tensor.to(self.device)
        
        with torch.no_grad():
            output = self.model(input_tensor)
        
        # Çıktıyı işle ve boyutlandır
        output = output.squeeze().permute(1, 2, 0).cpu().numpy()
        output = cv2.resize(output, (image.size[0] * scale_factor, image.size[1] * scale_factor))
        return Image.fromarray(np.uint8(output))

class SmartObjectDetection:
    def __init__(self):
        self.model = ResNet50(weights='imagenet')
        
    def detect_objects(self, image):
        """Görüntüdeki nesneleri tespit et"""
        img_array = np.array(image)
        img_resized = cv2.resize(img_array, (224, 224))
        img_expanded = np.expand_dims(img_resized, axis=0)
        img_preprocessed = tf.keras.applications.resnet50.preprocess_input(img_expanded)
        
        predictions = self.model.predict(img_preprocessed)
        return tf.keras.applications.resnet50.decode_predictions(predictions, top=5)[0]

class DeepBackgroundRemoval:
    def __init__(self):
        # U-Net benzeri bir model kullanacağız
        self.model = self._create_unet()
        
    def _create_unet(self):
        """Basit bir U-Net modeli oluştur"""
        inputs = tf.keras.layers.Input((None, None, 3))
        # Encoder
        conv1 = tf.keras.layers.Conv2D(64, 3, activation='relu', padding='same')(inputs)
        pool1 = tf.keras.layers.MaxPooling2D(pool_size=(2, 2))(conv1)
        
        # Decoder
        up1 = tf.keras.layers.Conv2DTranspose(64, 2, strides=(2,2), padding='same')(pool1)
        outputs = tf.keras.layers.Conv2D(1, 1, activation='sigmoid')(up1)
        
        return Model(inputs=[inputs], outputs=[outputs])
    
    def remove_background(self, image):
        """Arka planı derin öğrenme ile kaldır"""
        img_array = np.array(image)
        input_tensor = np.expand_dims(img_array, axis=0)
        
        mask = self.model.predict(input_tensor)[0]
        mask = cv2.resize(mask, (image.size[0], image.size[1]))
        
        # Maskeyi uygula
        result = img_array.copy()
        result[mask < 0.5] = [0, 0, 0, 0]
        return Image.fromarray(result)

class FaceEnhancement:
    def __init__(self):
        self.face_cascade = cv2.CascadeClassifier(cv2.data.haarcascades + 'haarcascade_frontalface_default.xml')
        self.landmark_detector = cv2.face.createFacemarkLBF()
        self.landmark_detector.loadModel('models/lbfmodel.yaml')
        
    def enhance_face(self, image, smoothing=1.0, brightness=1.0):
        """Gelişmiş yüz düzenleme"""
        img_array = np.array(image)
        gray = cv2.cvtColor(img_array, cv2.COLOR_RGB2GRAY)
        faces = self.face_cascade.detectMultiScale(gray, 1.3, 5)
        
        for (x, y, w, h) in faces:
            face_roi = img_array[y:y+h, x:x+w]
            
            # Yüz özelliklerini belirle
            landmarks = self.landmark_detector.fit(gray[y:y+h, x:x+w])
            
            # Cildi yumuşat
            face_roi = cv2.bilateralFilter(face_roi, 9, 75, 75)
            
            # Parlaklık ayarla
            face_roi = cv2.convertScaleAbs(face_roi, alpha=brightness, beta=10)
            
            img_array[y:y+h, x:x+w] = face_roi
            
        return Image.fromarray(img_array)

class SemanticSegmentation:
    def __init__(self):
        self.model = self._create_segmentation_model()
        
    def _create_segmentation_model(self):
        """DeepLab benzeri bir model oluştur"""
        base_model = VGG16(weights='imagenet', include_top=False)
        x = base_model.output
        x = tf.keras.layers.Conv2D(21, (1, 1), activation='softmax')(x)
        return Model(inputs=base_model.input, outputs=x)
    
    def segment_image(self, image):
        """Görüntüyü semantik olarak bölümle"""
        img_array = np.array(image)
        input_tensor = cv2.resize(img_array, (224, 224))
        input_tensor = np.expand_dims(input_tensor, axis=0)
        
        segmentation_map = self.model.predict(input_tensor)[0]
        segmentation_map = cv2.resize(segmentation_map, (image.size[0], image.size[1]))
        
        return Image.fromarray(np.uint8(segmentation_map * 255)) 