requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
scikit-image>=0.18.2
tensorflow>=2.6.0
matplotlib>=3.4.3
streamlit>=0.88.0
rembg>=2.0.0
face-recognition>=1.3.0
scipy>=1.7.0
dlib>=19.22.0
scikit-learn>=0.24.2
requests>=2.26.0
flask>=2.0.1
psutil>=5.8.0
python-dotenv>=0.19.0
dropbox>=11.0.0
google-auth>=2.3.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.2
tweepy>=4.0.0
instagram-private-api>=1.6.0
python-multipart>=0.0.5 